Fixed issue #1
This was caused by the fact that the output array might contain a reference to memory within the leftovers buffer. Previously this buffer would be updated before the output was pushed, causing data corruption.

I'm not exactly sure why this issue only happened at lower chunk sizes, maybe it was just a coincidence that it worked with chunk sizes that were multiples of 1024.